Questionable "optimization" in `eparens-osc.dart`.
The documentation for `IOSink#add` says not to modify the buffer after
it's been given as a parameter. I'm just gonna guess that it's
actually okay to modify it since I call `await stdout.flush()`, but I
don't actually know if this is true. It seems silly to recreate that
buffer on every single loop iteration though.