Sum/product/exponential primitives in Lisp, `eparens-repl`.
Previously I planned to make a type separate from `Procedure`, a DSL
for "combinatory" programs that e.g. ease analysis compared to
fexprs. Now I think I can identify this as a subset of procedures with
partial evaluation.

These are based on open deduction, I'll need to introduce that in
documentation later.