Remove combinators.
I'm going to refactor the interpreter to continuation passing style;
this is complicated in the presence of combinators. Specifically, for
products, how do you manage the continuation? Continuations serialize
the computation, but products are parallel. If you're forced to
implement the product combinator with `vau` expressions, then you're
also forced to serialize the computation. Whereas, it's strange for
the interpreter to serialize a primitive that claims to be the product
combinator.

Maybe there was a sensible thing to do here, but for now it seems
simpler to stick to `vau` expressions and push those sorts of
decisions to library code.