Detailed printing for error objects.
There's a couple bugs lurking in the reader (like just putting an
ending paren), and `define` doesn't really make sense just working on
the top level. I actually want to remove it entirely now, and use an
Awelon-like transaction model. Previously I thought it was important
to have algorithmic definitions, but now I realize you can model the
same with a single symbol that creates a record with many functions in
it, like "interface passing style".